IVGCVSW-8159 Fixed issues building with NDK r26
* The compiler shipped with NDK r26 has stricter rules around certain
  warnings and deprecation notices.
  * Fixed warnings for unqualified call to 'std::move'
  * Fixed error where the half values weren't being cast to a float
    when calling 'std::nan'
  * Removed unnecessary subtensor unit tests for neon

+++ include/armnnTestUtils/TensorHelpers.hpp
@@ -47,7 +47,7 @@ struct SelectiveComparer<T, false>
             return true;
         }
 
-        if (std::isnan(a) && std::isnan(b))
+        if (std::isnan(static_cast<float>(a)) && std::isnan(static_cast<float>(b)))
         {
             return true;
         }
